// angular/src/app/core/models/advisor/progress-vehicle.model.ts
export interface ProgressVehicleModel {
  carownername: string;
  cusType: number;
  cuswait: string;
  dlrId: number;
  engineno: string;
  id: number;
  name: string;
  registerno: string;
  repairorderno: string;
  wpId: number;
  estimateTime: string;
  state: number;
  carDeliveryTime: number;
  appId: number;
}